move tests to their own file
structured-clone.test.ts has 1.5GB-ArrayBuffer cases that run ~5s under
ASAN and intermittently time out; keep the new cases isolated so the
file-level pass/fail reflects only this change.
